+++ frontend/src/lib/api.js
@@ -146,7 +146,8 @@ export const validatorsAPI = {
   },
   getMyValidatorWallets: () => api.get('/validators/my-wallets/'),
   linkValidatorWalletsByOperator: (operatorAddress) => api.post('/validators/link-by-operator/', { operator_address: operatorAddress }),
-  getNetworks: () => api.get('/validators/wallets/networks/')
+  getNetworks: () => api.get('/validators/wallets/networks/'),
+  getWalletStats: () => api.get('/validators/wallets/stats/'),
 };
 
 // Builders API